Split out updatePostMetadata
